+++ NEWS.md
@@ -20,3 +20,8 @@ function that can only be used interactively.
 * Minor enhancement of type coercion to support more consistent behaviour for
 finding custom `type.XX` functions in downstream packages.
 * Funding statement added to README
+
+# interfacer 0.3.0
+
+* Consistency checking and recycling for non dataframe parameters
+* More consistent `type.XX` function behaviour